Refactor: pay 타입 분리
