+++ types/responses/pay.ts
@@ -1,0 +1,8 @@
+export interface PaymentListOutput {
+  payments: PaymentType[];
+}
+
+export type PaymentType = {
+  paymentId: number;
+  paymentName: string;
+};
